Exception handling - delete

+++ src/main/java/com/devRodrigoJr/courseSpringBootJPA/services/UserServices.java
@@ -2,12 +2,13 @@ package com.devRodrigoJr.courseSpringBootJPA.services;
 
 import com.devRodrigoJr.courseSpringBootJPA.entities.User;
 import com.devRodrigoJr.courseSpringBootJPA.repositories.UserRepository;
+import com.devRodrigoJr.courseSpringBootJPA.services.exceptions.DataBaseException;
 import com.devRodrigoJr.courseSpringBootJPA.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class UserServices {
@@ -29,7 +30,15 @@ public class UserServices {
     }
 
     public void delete(Long id) {
-        userRepository.deleteById(id);
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DataBaseException(e.getMessage());
+        }
+
     }
 
     public User update(Long id, User obj) {
